--- Spotifly/SpotifyAPI.swift
@@ -1,3363 +1,0 @@
-//
-//  SpotifyAPI.swift
-//  Spotifly
-//
-//  Spotify Web API client for fetching track metadata
-//
-
-import Foundation
-import os.log
-
-private let apiLogger = Logger(subsystem: "com.spotifly.app", category: "SpotifyAPI")
-
-// MARK: - Duration Formatting Protocol
-
-/// Protocol for types that have a total duration in milliseconds
-protocol DurationFormattable {
-    var totalDurationMs: Int? { get }
-}
-
-extension DurationFormattable {
-    /// Formats the total duration as "X hr Y min" or "Y min"
-    var formattedDuration: String? {
-        guard let totalDurationMs else { return nil }
-        let totalSeconds = totalDurationMs / 1000
-        let hours = totalSeconds / 3600
-        let minutes = (totalSeconds % 3600) / 60
-
-        if hours > 0 {
-            return String(format: "%d hr %d min", hours, minutes)
-        } else {
-            return String(format: "%d min", minutes)
-        }
-    }
-}
-
-// MARK: - Track Metadata
-
-/// Track metadata from Spotify
-struct TrackMetadata: Sendable {
-    let id: String
-    let name: String
-    let artistName: String
-    let albumName: String
-    let albumImageURL: URL?
-    let durationMs: Int
-    let previewURL: URL?
-
-    var durationFormatted: String {
-        let totalSeconds = durationMs / 1000
-        let minutes = totalSeconds / 60
-        let seconds = totalSeconds % 60
-        return String(format: "%d:%02d", minutes, seconds)
-    }
-}
-
-/// Playlist metadata from Spotify API
-struct APIPlaylist: Sendable, Identifiable, DurationFormattable {
-    let id: String
-    var name: String
-    let uri: String
-    let description: String?
-    let imageURL: URL?
-    var trackCount: Int
-    let ownerId: String
-    let ownerName: String
-    let isPublic: Bool?
-    let totalDurationMs: Int?
-}
-
-/// Response wrapper for playlists endpoint
-struct PlaylistsResponse: Sendable {
-    let playlists: [APIPlaylist]
-    let total: Int
-    let hasMore: Bool
-    let nextOffset: Int?
-}
-
-/// Album metadata from Spotify API
-struct APIAlbum: Sendable, Identifiable, DurationFormattable {
-    let id: String
-    let name: String
-    let uri: String
-    let artistName: String
-    let artistId: String?
-    let imageURL: URL?
-    let trackCount: Int
-    let releaseDate: String
-    let albumType: String?
-    let totalDurationMs: Int?
-    let externalUrl: String?
-}
-
-/// Response wrapper for albums endpoint
-struct AlbumsResponse: Sendable {
-    let albums: [APIAlbum]
-    let total: Int
-    let hasMore: Bool
-    let nextOffset: Int?
-}
-
-/// Artist metadata from Spotify API
-struct APIArtist: Sendable, Identifiable {
-    let id: String
-    let name: String
-    let uri: String
-    let imageURL: URL?
-    let genres: [String]
-    let followers: Int
-}
-
-/// Response wrapper for artists endpoint
-struct ArtistsResponse: Sendable {
-    let artists: [APIArtist]
-    let total: Int
-    let hasMore: Bool
-    let nextCursor: String?
-}
-
-/// Time range for top items (artists/tracks)
-enum TopItemsTimeRange: String, Sendable {
-    case shortTerm = "short_term" // ~4 weeks
-    case mediumTerm = "medium_term" // ~6 months (default)
-    case longTerm = "long_term" // ~1 year
-}
-
-/// Response wrapper for user's top artists endpoint
-struct TopArtistsResponse: Sendable {
-    let artists: [APIArtist]
-    let total: Int
-    let hasMore: Bool
-    let nextOffset: Int?
-}
-
-/// Response wrapper for new releases endpoint
-struct NewReleasesResponse: Sendable {
-    let albums: [APIAlbum]
-    let total: Int
-    let hasMore: Bool
-    let nextOffset: Int?
-}
-
-/// Saved track (favorite) metadata from Spotify
-struct SavedTrack: Sendable, Identifiable {
-    let id: String
-    let name: String
-    let artistName: String
-    let albumName: String
-    let imageURL: URL?
-    let durationMs: Int
-    let uri: String
-    let addedAt: String
-    let albumId: String?
-    let artistId: String?
-    let externalUrl: String?
-}
-
-/// Response wrapper for saved tracks endpoint
-struct SavedTracksResponse: Sendable {
-    let tracks: [SavedTrack]
-    let total: Int
-    let hasMore: Bool
-    let nextOffset: Int?
-}
-
-/// Album track (from album tracks endpoint)
-struct AlbumTrack: Sendable, Identifiable {
-    let id: String
-    let name: String
-    let uri: String
-    let artistName: String
-    let durationMs: Int
-    let trackNumber: Int
-    let artistId: String?
-    let externalUrl: String?
-}
-
-/// Playlist track (from playlist tracks endpoint)
-struct PlaylistTrack: Sendable, Identifiable {
-    let id: String
-    let name: String
-    let uri: String
-    let artistName: String
-    let albumName: String
-    let imageURL: URL?
-    let durationMs: Int
-    let addedAt: String
-    let albumId: String?
-    let artistId: String?
-    let externalUrl: String?
-}
-
-/// Search result type
-enum SearchType: String, Sendable {
-    case track
-    case album
-    case artist
-    case playlist
-}
-
-/// Track search result
-struct SearchTrack: Sendable, Identifiable {
-    let id: String
-    let name: String
-    let uri: String
-    let artistName: String
-    let albumName: String
-    let imageURL: URL?
-    let durationMs: Int
-    let albumId: String?
-    let artistId: String?
-    let externalUrl: String? // Web URL from Spotify API
-}
-
-/// Search results wrapper (uses unified Entity types)
-struct SearchResults: Sendable {
-    let tracks: [Track]
-    let albums: [Album]
-    let artists: [Artist]
-    let playlists: [Playlist]
-}
-
-/// Recently played context
-struct PlaybackContext: Sendable {
-    let type: String // "album", "playlist", "artist"
-    let uri: String
-}
-
-/// Recently played item
-struct RecentlyPlayedItem: Sendable, Identifiable {
-    let id: String // Use played_at as ID since tracks can be played multiple times
-    let track: SearchTrack
-    let playedAt: String
-    let context: PlaybackContext?
-}
-
-/// Recently played response wrapper
-struct RecentlyPlayedResponse: Sendable {
-    let items: [RecentlyPlayedItem]
-}
-
-/// Spotify Connect device
-struct SpotifyDevice: Sendable, Identifiable {
-    let id: String
-    let name: String
-    let type: String // "Computer", "Smartphone", "Speaker", etc.
-    let isActive: Bool
-    let isPrivateSession: Bool
-    let isRestricted: Bool
-    let volumePercent: Int?
-}
-
-/// Devices response wrapper
-struct DevicesResponse: Sendable {
-    let devices: [SpotifyDevice]
-}
-
-/// Track info in playback state
-struct PlaybackTrack: Sendable, Identifiable {
-    let id: String
-    let name: String
-    let uri: String
-    let artistName: String
-    let albumName: String
-    let imageURL: URL?
-    let durationMs: Int
-
-    /// Parse from Spotify API JSON
-    init?(from json: [String: Any]) {
-        guard let id = json["id"] as? String,
-              let name = json["name"] as? String,
-              let uri = json["uri"] as? String,
-              let durationMs = json["duration_ms"] as? Int
-        else {
-            return nil
-        }
-
-        self.id = id
-        self.name = name
-        self.uri = uri
-        self.durationMs = durationMs
-
-        let artistsArray = json["artists"] as? [[String: Any]]
-        artistName = artistsArray?.compactMap { $0["name"] as? String }.joined(separator: ", ") ?? "Unknown"
-
-        let albumData = json["album"] as? [String: Any]
-        albumName = albumData?["name"] as? String ?? ""
-        let albumImages = albumData?["images"] as? [[String: Any]]
-        let imageURLString = albumImages?.first?["url"] as? String
-        imageURL = imageURLString.flatMap { URL(string: $0) }
-    }
-}
-
-/// Current playback state from Spotify
-struct PlaybackState: Sendable {
-    let device: SpotifyDevice?
-    let isPlaying: Bool
-    let progressMs: Int
-    let currentTrack: PlaybackTrack?
-    let shuffleState: Bool
-    let repeatState: String
-}
-
-/// Track in the playback queue (same structure as PlaybackTrack)
-typealias QueueTrack = PlaybackTrack
-
-/// Queue response from Spotify
-struct QueueResponse: Sendable {
-    let currentlyPlaying: QueueTrack?
-    let queue: [QueueTrack]
-}
-
-/// Errors from Spotify API
-enum SpotifyAPIError: Error, LocalizedError {
-    case invalidURI
-    case networkError(Error)
-    case invalidResponse
-    case unauthorized
-    case notFound
-    case apiError(String)
-
-    var errorDescription: String? {
-        switch self {
-        case .invalidURI:
-            "Invalid Spotify URI format"
-        case let .networkError(error):
-            "Network error: \(error.localizedDescription)"
-        case .invalidResponse:
-            "Invalid response from Spotify"
-        case .unauthorized:
-            "Unauthorized - please log in again"
-        case .notFound:
-            "Track not found"
-        case let .apiError(message):
-            "Spotify API error: \(message)"
-        }
-    }
-}
-
-/// Spotify Web API client
-enum SpotifyAPI {
-    private static let baseURL = "https://api.spotify.com/v1"
-
-    /// Parses a Spotify URI (spotify:track:xxx) and returns the track ID
-    static func parseTrackURI(_ uri: String) -> String? {
-        let trimmed = uri.trimmingCharacters(in: .whitespacesAndNewlines)
-
-        // Handle spotify:track:ID format
-        if trimmed.hasPrefix("spotify:track:") {
-            return String(trimmed.dropFirst("spotify:track:".count))
-        }
-
-        // Handle https://open.spotify.com/track/ID format
-        if let url = URL(string: trimmed),
-           url.host == "open.spotify.com",
-           url.pathComponents.count >= 3,
-           url.pathComponents[1] == "track"
-        {
-            // Remove any query parameters from the track ID
-            return url.pathComponents[2].components(separatedBy: "?").first
-        }
-
-        // If it looks like just an ID (22 chars, alphanumeric)
-        if trimmed.count == 22, trimmed.allSatisfy({ $0.isLetter || $0.isNumber }) {
-            return trimmed
-        }
-
-        return nil
-    }
-
-    /// Fetches track metadata from Spotify Web API
-    static func fetchTrackMetadata(trackId: String, accessToken: String) async throws -> TrackMetadata {
-        let urlString = "\(baseURL)/tracks/\(trackId)?fields=id,name,duration_ms,artists(name),album(name,images),preview_url"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Accept")
-
-        let (data, response): (Data, URLResponse)
-        do {
-            (data, response) = try await URLSession.shared.data(for: request)
-        } catch {
-            throw SpotifyAPIError.networkError(error)
-        }
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-
-        // Parse JSON response
-        guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        guard let id = json["id"] as? String,
-              let name = json["name"] as? String,
-              let durationMs = json["duration_ms"] as? Int
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        // Extract artist name(s)
-        var artistName = "Unknown Artist"
-        if let artists = json["artists"] as? [[String: Any]] {
-            let artistNames = artists.compactMap { $0["name"] as? String }
-            artistName = artistNames.joined(separator: ", ")
-        }
-
-        // Extract album info
-        var albumName = "Unknown Album"
-        var albumImageURL: URL? = nil
-        if let album = json["album"] as? [String: Any] {
-            albumName = album["name"] as? String ?? "Unknown Album"
-
-            if let images = album["images"] as? [[String: Any]],
-               let firstImage = images.first,
-               let urlString = firstImage["url"] as? String
-            {
-                albumImageURL = URL(string: urlString)
-            }
-        }
-
-        // Extract preview URL
-        var previewURL: URL? = nil
-        if let previewUrlString = json["preview_url"] as? String {
-            previewURL = URL(string: previewUrlString)
-        }
-
-        return TrackMetadata(
-            id: id,
-            name: name,
-            artistName: artistName,
-            albumName: albumName,
-            albumImageURL: albumImageURL,
-            durationMs: durationMs,
-            previewURL: previewURL,
-        )
-    }
-
-    /// Fetches user's playlists from Spotify Web API
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - limit: Number of playlists to fetch (max 50)
-    ///   - offset: Offset for pagination
-    static func fetchUserPlaylists(accessToken: String, limit: Int = 50, offset: Int = 0) async throws -> PlaylistsResponse {
-        let urlString = "\(baseURL)/me/playlists?limit=\(limit)&offset=\(offset)&fields=items(id,name,uri,description,images,tracks(total,items(track(duration_ms))),public,owner(id,display_name)),total,next"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Accept")
-
-        let (data, response): (Data, URLResponse)
-        do {
-            (data, response) = try await URLSession.shared.data(for: request)
-        } catch {
-            throw SpotifyAPIError.networkError(error)
-        }
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-
-        // Parse JSON response
-        guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        guard let items = json["items"] as? [[String: Any]],
-              let total = json["total"] as? Int
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        let playlists = items.compactMap { item -> APIPlaylist? in
-            guard let id = item["id"] as? String,
-                  let name = item["name"] as? String,
-                  let uri = item["uri"] as? String,
-                  let tracks = item["tracks"] as? [String: Any],
-                  let trackCount = tracks["total"] as? Int,
-                  let owner = item["owner"] as? [String: Any],
-                  let ownerId = owner["id"] as? String
-            else {
-                return nil
-            }
-
-            let ownerName = owner["display_name"] as? String ?? ownerId
-            let description = item["description"] as? String
-            let isPublic = item["public"] as? Bool ?? false
-
-            var imageURL: URL?
-            if let images = item["images"] as? [[String: Any]],
-               let firstImage = images.first,
-               let urlString = firstImage["url"] as? String
-            {
-                imageURL = URL(string: urlString)
-            }
-
-            // Calculate total duration from tracks
-            var totalDurationMs: Int?
-            if let items = tracks["items"] as? [[String: Any]] {
-                let durations = items.compactMap { item -> Int? in
-                    guard let track = item["track"] as? [String: Any],
-                          let duration = track["duration_ms"] as? Int
-                    else {
-                        return nil
-                    }
-                    return duration
-                }
-                if !durations.isEmpty {
-                    totalDurationMs = durations.reduce(0, +)
-                }
-            }
-
-            return APIPlaylist(
-                id: id,
-                name: name,
-                uri: uri,
-                description: description,
-                imageURL: imageURL,
-                trackCount: trackCount,
-                ownerId: ownerId,
-                ownerName: ownerName,
-                isPublic: isPublic,
-                totalDurationMs: totalDurationMs,
-            )
-        }
-
-        let next = json["next"] as? String
-        let hasMore = next != nil
-        let nextOffset = hasMore ? offset + limit : nil
-
-        return PlaylistsResponse(
-            playlists: playlists,
-            total: total,
-            hasMore: hasMore,
-            nextOffset: nextOffset,
-        )
-    }
-
-    /// Fetches a single playlist's details from Spotify Web API
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - playlistId: Playlist ID
-    static func fetchPlaylistDetails(accessToken: String, playlistId: String) async throws -> APIPlaylist {
-        // Add fields parameter to request only what we need, and market=from_token for region-specific playlists
-        let urlString = "\(baseURL)/playlists/\(playlistId)?fields=id,name,description,images,tracks(total,items(track(duration_ms))),uri,public,owner(id,display_name)&market=from_token"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Accept")
-
-        let (data, response): (Data, URLResponse)
-        do {
-            (data, response) = try await URLSession.shared.data(for: request)
-        } catch {
-            throw SpotifyAPIError.networkError(error)
-        }
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-
-        // Parse JSON response
-        guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        guard let id = json["id"] as? String,
-              let name = json["name"] as? String,
-              let uri = json["uri"] as? String,
-              let tracks = json["tracks"] as? [String: Any],
-              let trackCount = tracks["total"] as? Int,
-              let owner = json["owner"] as? [String: Any],
-              let ownerId = owner["id"] as? String
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        // Owner's display_name can be null, fall back to id
-        let ownerName = owner["display_name"] as? String ?? ownerId
-
-        let description = json["description"] as? String
-
-        var imageURL: URL?
-        if let images = json["images"] as? [[String: Any]],
-           let firstImage = images.first,
-           let urlString = firstImage["url"] as? String
-        {
-            imageURL = URL(string: urlString)
-        }
-
-        // Calculate total duration from tracks
-        var totalDurationMs: Int?
-        if let items = tracks["items"] as? [[String: Any]] {
-            let durations = items.compactMap { item -> Int? in
-                guard let track = item["track"] as? [String: Any],
-                      let duration = track["duration_ms"] as? Int
-                else {
-                    return nil
-                }
-                return duration
-            }
-            if !durations.isEmpty {
-                totalDurationMs = durations.reduce(0, +)
-            }
-        }
-
-        return APIPlaylist(
-            id: id,
-            name: name,
-            uri: uri,
-            description: description,
-            imageURL: imageURL,
-            trackCount: trackCount,
-            ownerId: ownerId,
-            ownerName: ownerName,
-            isPublic: nil,
-            totalDurationMs: totalDurationMs,
-        )
-    }
-
-    /// Fetches a single album's details from Spotify Web API
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - albumId: Album ID
-    static func fetchAlbumDetails(accessToken: String, albumId: String) async throws -> APIAlbum {
-        let urlString = "\(baseURL)/albums/\(albumId)?fields=id,name,uri,total_tracks,release_date,artists(id,name),images,tracks(items(duration_ms)),external_urls(spotify)"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Accept")
-
-        let (data, response): (Data, URLResponse)
-        do {
-            (data, response) = try await URLSession.shared.data(for: request)
-        } catch {
-            throw SpotifyAPIError.networkError(error)
-        }
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-
-        // Parse JSON response
-        guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        guard let id = json["id"] as? String,
-              let name = json["name"] as? String,
-              let uri = json["uri"] as? String,
-              let totalTracks = json["total_tracks"] as? Int,
-              let releaseDate = json["release_date"] as? String,
-              let artists = json["artists"] as? [[String: Any]]
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        let artistName = artists.first?["name"] as? String ?? "Unknown Artist"
-        let artistId = artists.first?["id"] as? String
-
-        var imageURL: URL?
-        if let images = json["images"] as? [[String: Any]],
-           let firstImage = images.first,
-           let urlString = firstImage["url"] as? String
-        {
-            imageURL = URL(string: urlString)
-        }
-
-        // Calculate total duration from tracks
-        var totalDurationMs: Int?
-        if let tracksObj = json["tracks"] as? [String: Any],
-           let items = tracksObj["items"] as? [[String: Any]]
-        {
-            let durations = items.compactMap { $0["duration_ms"] as? Int }
-            if !durations.isEmpty {
-                totalDurationMs = durations.reduce(0, +)
-            }
-        }
-
-        // Parse external URL
-        let externalUrls = json["external_urls"] as? [String: Any]
-        let externalUrl = externalUrls?["spotify"] as? String
-
-        return APIAlbum(
-            id: id,
-            name: name,
-            uri: uri,
-            artistName: artistName,
-            artistId: artistId,
-            imageURL: imageURL,
-            trackCount: totalTracks,
-            releaseDate: releaseDate,
-            albumType: nil,
-            totalDurationMs: totalDurationMs,
-            externalUrl: externalUrl,
-        )
-    }
-
-    /// Fetches a single artist's details from Spotify Web API
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - artistId: Artist ID
-    static func fetchArtistDetails(accessToken: String, artistId: String) async throws -> APIArtist {
-        let urlString = "\(baseURL)/artists/\(artistId)?fields=id,name,uri,genres,followers(total),images"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Accept")
-
-        let (data, response): (Data, URLResponse)
-        do {
-            (data, response) = try await URLSession.shared.data(for: request)
-        } catch {
-            throw SpotifyAPIError.networkError(error)
-        }
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-
-        // Parse JSON response
-        guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        guard let id = json["id"] as? String,
-              let name = json["name"] as? String,
-              let uri = json["uri"] as? String
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        let genres = json["genres"] as? [String] ?? []
-        let followers = (json["followers"] as? [String: Any])?["total"] as? Int ?? 0
-
-        var imageURL: URL?
-        if let images = json["images"] as? [[String: Any]],
-           let firstImage = images.first,
-           let urlString = firstImage["url"] as? String
-        {
-            imageURL = URL(string: urlString)
-        }
-
-        return APIArtist(
-            id: id,
-            name: name,
-            uri: uri,
-            imageURL: imageURL,
-            genres: genres,
-            followers: followers,
-        )
-    }
-
-    /// Fetches user's saved albums from Spotify Web API
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - limit: Number of albums to fetch (max 50)
-    ///   - offset: Offset for pagination
-    static func fetchUserAlbums(accessToken: String, limit: Int = 50, offset: Int = 0) async throws -> AlbumsResponse {
-        let urlString = "\(baseURL)/me/albums?limit=\(limit)&offset=\(offset)&fields=items(album(id,name,uri,total_tracks,release_date,album_type,artists(name),images,tracks(items(duration_ms)))),total,next"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Accept")
-
-        let (data, response): (Data, URLResponse)
-        do {
-            (data, response) = try await URLSession.shared.data(for: request)
-        } catch {
-            throw SpotifyAPIError.networkError(error)
-        }
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-
-        // Parse JSON response
-        guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        guard let items = json["items"] as? [[String: Any]],
-              let total = json["total"] as? Int
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        let albums = items.compactMap { item -> APIAlbum? in
-            guard let album = item["album"] as? [String: Any],
-                  let id = album["id"] as? String,
-                  let name = album["name"] as? String,
-                  let uri = album["uri"] as? String,
-                  let totalTracks = album["total_tracks"] as? Int,
-                  let releaseDate = album["release_date"] as? String,
-                  let albumType = album["album_type"] as? String
-            else {
-                return nil
-            }
-
-            var artistName = "Unknown Artist"
-            if let artists = album["artists"] as? [[String: Any]] {
-                let artistNames = artists.compactMap { $0["name"] as? String }
-                artistName = artistNames.joined(separator: ", ")
-            }
-
-            var imageURL: URL?
-            if let images = album["images"] as? [[String: Any]],
-               let firstImage = images.first,
-               let urlString = firstImage["url"] as? String
-            {
-                imageURL = URL(string: urlString)
-            }
-
-            // Calculate total duration from tracks
-            var totalDurationMs: Int?
-            if let tracksObj = album["tracks"] as? [String: Any],
-               let items = tracksObj["items"] as? [[String: Any]]
-            {
-                let durations = items.compactMap { $0["duration_ms"] as? Int }
-                if !durations.isEmpty {
-                    totalDurationMs = durations.reduce(0, +)
-                }
-            }
-
-            return APIAlbum(
-                id: id,
-                name: name,
-                uri: uri,
-                artistName: artistName,
-                artistId: nil,
-                imageURL: imageURL,
-                trackCount: totalTracks,
-                releaseDate: releaseDate,
-                albumType: albumType,
-                totalDurationMs: totalDurationMs,
-                externalUrl: nil,
-            )
-        }
-
-        let next = json["next"] as? String
-        let hasMore = next != nil
-        let nextOffset = hasMore ? offset + limit : nil
-
-        return AlbumsResponse(
-            albums: albums,
-            total: total,
-            hasMore: hasMore,
-            nextOffset: nextOffset,
-        )
-    }
-
-    /// Fetches user's followed artists from Spotify Web API
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - limit: Number of artists to fetch (max 50)
-    ///   - after: Cursor for pagination (artist ID)
-    static func fetchUserArtists(accessToken: String, limit: Int = 50, after: String? = nil) async throws -> ArtistsResponse {
-        var urlString = "\(baseURL)/me/following?type=artist&limit=\(limit)&fields=artists(items(id,name,uri,genres,followers(total),images),total,cursors)"
-        if let after {
-            urlString += "&after=\(after)"
-        }
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Accept")
-
-        let (data, response): (Data, URLResponse)
-        do {
-            (data, response) = try await URLSession.shared.data(for: request)
-        } catch {
-            throw SpotifyAPIError.networkError(error)
-        }
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-
-        // Parse JSON response
-        guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-              let artistsContainer = json["artists"] as? [String: Any],
-              let items = artistsContainer["items"] as? [[String: Any]],
-              let total = artistsContainer["total"] as? Int
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        let artists = items.compactMap { item -> APIArtist? in
-            guard let id = item["id"] as? String,
-                  let name = item["name"] as? String,
-                  let uri = item["uri"] as? String
-            else {
-                return nil
-            }
-
-            let genres = (item["genres"] as? [String]) ?? []
-            let followersDict = item["followers"] as? [String: Any]
-            let followers = followersDict?["total"] as? Int ?? 0
-
-            var imageURL: URL?
-            if let images = item["images"] as? [[String: Any]],
-               let firstImage = images.first,
-               let urlString = firstImage["url"] as? String
-            {
-                imageURL = URL(string: urlString)
-            }
-
-            return APIArtist(
-                id: id,
-                name: name,
-                uri: uri,
-                imageURL: imageURL,
-                genres: genres,
-                followers: followers
-            )
-        }
-
-        let cursors = artistsContainer["cursors"] as? [String: Any]
-        let afterCursor = cursors?["after"] as? String
-
-        return ArtistsResponse(
-            artists: artists,
-            total: total,
-            hasMore: afterCursor != nil,
-            nextCursor: afterCursor,
-        )
-    }
-
-    /// Fetches user's saved tracks (favorites) from Spotify Web API
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - limit: Number of tracks to fetch (max 50)
-    ///   - offset: Offset for pagination
-    static func fetchUserSavedTracks(accessToken: String, limit: Int = 50, offset: Int = 0) async throws -> SavedTracksResponse {
-        let urlString = "\(baseURL)/me/tracks?limit=\(limit)&offset=\(offset)&fields=items(track(id,name,uri,duration_ms,artists(id,name),album(id,name,images),external_urls(spotify)),added_at),total,next"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Accept")
-
-        let (data, response): (Data, URLResponse)
-        do {
-            (data, response) = try await URLSession.shared.data(for: request)
-        } catch {
-            throw SpotifyAPIError.networkError(error)
-        }
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-
-        // Parse JSON response
-        guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        guard let items = json["items"] as? [[String: Any]],
-              let total = json["total"] as? Int
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        let tracks = items.compactMap { item -> SavedTrack? in
-            guard let track = item["track"] as? [String: Any],
-                  let id = track["id"] as? String,
-                  let name = track["name"] as? String,
-                  let uri = track["uri"] as? String,
-                  let durationMs = track["duration_ms"] as? Int,
-                  let addedAt = item["added_at"] as? String
-            else {
-                return nil
-            }
-
-            var artistName = "Unknown Artist"
-            var artistId: String?
-            if let artists = track["artists"] as? [[String: Any]] {
-                let artistNames = artists.compactMap { $0["name"] as? String }
-                artistName = artistNames.joined(separator: ", ")
-                artistId = artists.first?["id"] as? String
-            }
-
-            var albumName = "Unknown Album"
-            var albumId: String?
-            var imageURL: URL?
-            if let album = track["album"] as? [String: Any] {
-                albumName = album["name"] as? String ?? "Unknown Album"
-                albumId = album["id"] as? String
-
-                if let images = album["images"] as? [[String: Any]],
-                   let firstImage = images.first,
-                   let urlString = firstImage["url"] as? String
-                {
-                    imageURL = URL(string: urlString)
-                }
-            }
-
-            let externalUrls = track["external_urls"] as? [String: Any]
-            let externalUrl = externalUrls?["spotify"] as? String
-
-            return SavedTrack(
-                id: id,
-                name: name,
-                artistName: artistName,
-                albumName: albumName,
-                imageURL: imageURL,
-                durationMs: durationMs,
-                uri: uri,
-                addedAt: addedAt,
-                albumId: albumId,
-                artistId: artistId,
-                externalUrl: externalUrl,
-            )
-        }
-
-        let next = json["next"] as? String
-        let hasMore = next != nil
-        let nextOffset = hasMore ? offset + limit : nil
-
-        return SavedTracksResponse(
-            tracks: tracks,
-            total: total,
-            hasMore: hasMore,
-            nextOffset: nextOffset,
-        )
-    }
-
-    /// Saves a track to user's library (favorites)
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - trackId: The Spotify ID of the track to save
-    static func saveTrack(accessToken: String, trackId: String) async throws {
-        let urlString = "\(baseURL)/me/tracks"
-        #if DEBUG
-            apiLogger.debug("[PUT] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-
-        // Create request body
-        let body = ["ids": [trackId]]
-        request.httpBody = try? JSONSerialization.data(withJSONObject: body)
-
-        let (data, response): (Data, URLResponse)
-        do {
-            (data, response) = try await URLSession.shared.data(for: request)
-        } catch {
-            throw SpotifyAPIError.networkError(error)
-        }
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200, 201:
-            // Success - track saved
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    /// Checks if a track is saved in user's library
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - trackId: The Spotify ID of the track to check
-    /// - Returns: True if track is saved, false otherwise
-    static func checkSavedTrack(accessToken: String, trackId: String) async throws -> Bool {
-        let result = try await checkSavedTracks(accessToken: accessToken, trackIds: [trackId])
-        return result[trackId] ?? false
-    }
-
-    /// Batch checks if multiple tracks are saved in user's library
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - trackIds: Array of Spotify track IDs to check (max 50 per request)
-    /// - Returns: Dictionary mapping track ID to saved status
-    static func checkSavedTracks(accessToken: String, trackIds: [String]) async throws -> [String: Bool] {
-        guard !trackIds.isEmpty else { return [:] }
-
-        // Spotify API allows max 50 IDs per request
-        let batchSize = 50
-        var results: [String: Bool] = [:]
-
-        for batchStart in stride(from: 0, to: trackIds.count, by: batchSize) {
-            let batchEnd = min(batchStart + batchSize, trackIds.count)
-            let batch = Array(trackIds[batchStart ..< batchEnd])
-            let idsParam = batch.joined(separator: ",")
-
-            let urlString = "\(baseURL)/me/tracks/contains?ids=\(idsParam)"
-            #if DEBUG
-                apiLogger.debug("[GET] \(urlString)")
-            #endif
-
-            guard let url = URL(string: urlString) else {
-                throw SpotifyAPIError.invalidURI
-            }
-
-            var request = URLRequest(url: url)
-            request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-            let (data, response): (Data, URLResponse)
-            do {
-                (data, response) = try await URLSession.shared.data(for: request)
-            } catch {
-                throw SpotifyAPIError.networkError(error)
-            }
-
-            guard let httpResponse = response as? HTTPURLResponse else {
-                throw SpotifyAPIError.invalidResponse
-            }
-
-            switch httpResponse.statusCode {
-            case 200:
-                guard let jsonArray = try? JSONSerialization.jsonObject(with: data) as? [Bool],
-                      jsonArray.count == batch.count
-                else {
-                    throw SpotifyAPIError.invalidResponse
-                }
-                // Map results back to track IDs
-                for (index, trackId) in batch.enumerated() {
-                    results[trackId] = jsonArray[index]
-                }
-            case 401:
-                throw SpotifyAPIError.unauthorized
-            default:
-                if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-                   let error = errorJson["error"] as? [String: Any],
-                   let message = error["message"] as? String
-                {
-                    throw SpotifyAPIError.apiError(message)
-                }
-                throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-            }
-        }
-
-        return results
-    }
-
-    /// Removes a track from user's saved tracks (unfavorites)
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - trackId: The Spotify ID of the track to remove
-    static func removeSavedTrack(accessToken: String, trackId: String) async throws {
-        let urlString = "\(baseURL)/me/tracks?ids=\(trackId)"
-        #if DEBUG
-            apiLogger.debug("[DELETE] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "DELETE"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Accept")
-
-        let (data, response): (Data, URLResponse)
-        do {
-            (data, response) = try await URLSession.shared.data(for: request)
-        } catch {
-            throw SpotifyAPIError.networkError(error)
-        }
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            // Success - track removed
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    // MARK: - Album Tracks
-
-    /// Fetches tracks for a specific album
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - albumId: The Spotify ID of the album
-    static func fetchAlbumTracks(accessToken: String, albumId: String) async throws -> [AlbumTrack] {
-        let urlString = "\(baseURL)/albums/\(albumId)/tracks?limit=50&fields=items(id,name,uri,duration_ms,track_number,artists(id,name),external_urls(spotify))"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-                  let items = json["items"] as? [[String: Any]]
-            else {
-                throw SpotifyAPIError.invalidResponse
-            }
-
-            let tracks = items.compactMap { item -> AlbumTrack? in
-                guard let id = item["id"] as? String,
-                      let name = item["name"] as? String,
-                      let uri = item["uri"] as? String,
-                      let durationMs = item["duration_ms"] as? Int,
-                      let trackNumber = item["track_number"] as? Int
-                else {
-                    return nil
-                }
-
-                let artistsArray = item["artists"] as? [[String: Any]]
-                let artistName = artistsArray?.first?["name"] as? String ?? "Unknown"
-                let artistId = artistsArray?.first?["id"] as? String
-
-                let externalUrls = item["external_urls"] as? [String: Any]
-                let externalUrl = externalUrls?["spotify"] as? String
-
-                return AlbumTrack(
-                    id: id,
-                    name: name,
-                    uri: uri,
-                    artistName: artistName,
-                    durationMs: durationMs,
-                    trackNumber: trackNumber,
-                    artistId: artistId,
-                    externalUrl: externalUrl,
-                )
-            }
-
-            return tracks
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    // MARK: - Playlist Tracks
-
-    /// Fetches tracks for a specific playlist
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - playlistId: The Spotify ID of the playlist
-    static func fetchPlaylistTracks(accessToken: String, playlistId: String) async throws -> [PlaylistTrack] {
-        let urlString = "\(baseURL)/playlists/\(playlistId)/tracks?limit=50&fields=items(track(id,name,uri,duration_ms,artists(id,name),album(id,name,images),external_urls(spotify)),added_at)"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-                  let items = json["items"] as? [[String: Any]]
-            else {
-                throw SpotifyAPIError.invalidResponse
-            }
-
-            let tracks = items.compactMap { item -> PlaylistTrack? in
-                guard let track = item["track"] as? [String: Any],
-                      let id = track["id"] as? String,
-                      let name = track["name"] as? String,
-                      let uri = track["uri"] as? String,
-                      let durationMs = track["duration_ms"] as? Int
-                else {
-                    return nil
-                }
-
-                let artistsArray = track["artists"] as? [[String: Any]]
-                let artistName = artistsArray?.first?["name"] as? String ?? "Unknown"
-                let artistId = artistsArray?.first?["id"] as? String
-
-                let albumData = track["album"] as? [String: Any]
-                let albumName = albumData?["name"] as? String ?? ""
-                let albumId = albumData?["id"] as? String
-                let albumImages = albumData?["images"] as? [[String: Any]]
-                let imageURLString = albumImages?.first?["url"] as? String
-                let imageURL = imageURLString.flatMap { URL(string: $0) }
-
-                let externalUrls = track["external_urls"] as? [String: Any]
-                let externalUrl = externalUrls?["spotify"] as? String
-
-                let addedAt = item["added_at"] as? String ?? ""
-
-                return PlaylistTrack(
-                    id: id,
-                    name: name,
-                    uri: uri,
-                    artistName: artistName,
-                    albumName: albumName,
-                    imageURL: imageURL,
-                    durationMs: durationMs,
-                    addedAt: addedAt,
-                    albumId: albumId,
-                    artistId: artistId,
-                    externalUrl: externalUrl,
-                )
-            }
-
-            return tracks
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    // MARK: - Artist Top Tracks
-
-    /// Fetches top tracks for a specific artist
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - artistId: The Spotify ID of the artist
-    static func fetchArtistTopTracks(accessToken: String, artistId: String) async throws -> [SearchTrack] {
-        // Use market from token to get region-specific results
-        // Note: top-tracks endpoint doesn't support fields parameter, returns full track objects
-        let urlString = "\(baseURL)/artists/\(artistId)/top-tracks?market=from_token"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-                  let tracks = json["tracks"] as? [[String: Any]]
-            else {
-                throw SpotifyAPIError.invalidResponse
-            }
-
-            let topTracks = tracks.compactMap { item -> SearchTrack? in
-                guard let id = item["id"] as? String,
-                      let name = item["name"] as? String,
-                      let uri = item["uri"] as? String,
-                      let durationMs = item["duration_ms"] as? Int
-                else {
-                    return nil
-                }
-
-                let artistsArray = item["artists"] as? [[String: Any]]
-                let artistName = artistsArray?.first?["name"] as? String ?? "Unknown"
-                let artistId = artistsArray?.first?["id"] as? String
-
-                let albumData = item["album"] as? [String: Any]
-                let albumName = albumData?["name"] as? String ?? ""
-                let albumId = albumData?["id"] as? String
-                let albumImages = albumData?["images"] as? [[String: Any]]
-                let imageURLString = albumImages?.first?["url"] as? String
-                let imageURL = imageURLString.flatMap { URL(string: $0) }
-
-                let externalUrls = item["external_urls"] as? [String: Any]
-                let externalUrl = externalUrls?["spotify"] as? String
-
-                return SearchTrack(
-                    id: id,
-                    name: name,
-                    uri: uri,
-                    artistName: artistName,
-                    albumName: albumName,
-                    imageURL: imageURL,
-                    durationMs: durationMs,
-                    albumId: albumId,
-                    artistId: artistId,
-                    externalUrl: externalUrl,
-                )
-            }
-
-            return topTracks
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    // MARK: - Artist Albums
-
-    /// Fetches albums for a specific artist
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - artistId: The Spotify ID of the artist
-    ///   - limit: Maximum number of albums to return (default 50)
-    /// - Returns: Array of albums by the artist
-    static func fetchArtistAlbums(
-        accessToken: String,
-        artistId: String,
-        limit: Int = 50,
-    ) async throws -> [APIAlbum] {
-        let urlString = "\(baseURL)/artists/\(artistId)/albums?include_groups=album,single&market=from_token&limit=\(limit)"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-                  let items = json["items"] as? [[String: Any]]
-            else {
-                throw SpotifyAPIError.invalidResponse
-            }
-
-            let albums = items.compactMap { item -> APIAlbum? in
-                guard let id = item["id"] as? String,
-                      let name = item["name"] as? String,
-                      let uri = item["uri"] as? String,
-                      let totalTracks = item["total_tracks"] as? Int,
-                      let releaseDate = item["release_date"] as? String
-                else {
-                    return nil
-                }
-
-                let artistsArray = item["artists"] as? [[String: Any]]
-                let artistName = artistsArray?.first?["name"] as? String ?? "Unknown"
-
-                let images = item["images"] as? [[String: Any]]
-                let imageURLString = images?.first?["url"] as? String
-                let imageURL = imageURLString.flatMap { URL(string: $0) }
-
-                return APIAlbum(
-                    id: id,
-                    name: name,
-                    uri: uri,
-                    artistName: artistName,
-                    artistId: nil,
-                    imageURL: imageURL,
-                    trackCount: totalTracks,
-                    releaseDate: releaseDate,
-                    albumType: nil,
-                    totalDurationMs: nil,
-                    externalUrl: nil,
-                )
-            }
-
-            return albums
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    // MARK: - Recommendations (Radio)
-
-    /// Fetches track recommendations based on a seed track (for "Start Radio" feature)
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - seedTrackId: The Spotify ID of the seed track
-    ///   - limit: Number of tracks to return (default 50, max 100)
-    /// - Returns: Array of recommended tracks
-    static func fetchRecommendations(
-        accessToken: String,
-        seedTrackId: String,
-        limit: Int = 50,
-    ) async throws -> [SearchTrack] {
-        let urlString = "\(baseURL)/recommendations?seed_tracks=\(seedTrackId)&limit=\(limit)"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-                  let tracks = json["tracks"] as? [[String: Any]]
-            else {
-                throw SpotifyAPIError.invalidResponse
-            }
-
-            let recommendedTracks = tracks.compactMap { item -> SearchTrack? in
-                guard let id = item["id"] as? String,
-                      let name = item["name"] as? String,
-                      let uri = item["uri"] as? String,
-                      let durationMs = item["duration_ms"] as? Int
-                else {
-                    return nil
-                }
-
-                let artistsArray = item["artists"] as? [[String: Any]]
-                let artistName = artistsArray?.first?["name"] as? String ?? "Unknown"
-                let artistId = artistsArray?.first?["id"] as? String
-
-                let albumData = item["album"] as? [String: Any]
-                let albumName = albumData?["name"] as? String ?? ""
-                let albumId = albumData?["id"] as? String
-                let albumImages = albumData?["images"] as? [[String: Any]]
-                let imageURLString = albumImages?.first?["url"] as? String
-                let imageURL = imageURLString.flatMap { URL(string: $0) }
-
-                let externalUrls = item["external_urls"] as? [String: Any]
-                let externalUrl = externalUrls?["spotify"] as? String
-
-                return SearchTrack(
-                    id: id,
-                    name: name,
-                    uri: uri,
-                    artistName: artistName,
-                    albumName: albumName,
-                    imageURL: imageURL,
-                    durationMs: durationMs,
-                    albumId: albumId,
-                    artistId: artistId,
-                    externalUrl: externalUrl,
-                )
-            }
-
-            return recommendedTracks
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            // Try to get actual error message from Spotify
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError("Recommendations API: \(message)")
-            }
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    // MARK: - Search
-
-    /// Performs a search across Spotify's catalog
-    static func search(
-        accessToken: String,
-        query: String,
-        types: [SearchType] = [.track, .album, .artist, .playlist],
-        limit: Int = 20,
-    ) async throws -> SearchResults {
-        let typesParam = types.map(\.rawValue).joined(separator: ",")
-        let encodedQuery = query.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed) ?? query
-
-        let urlString = "\(baseURL)/search?q=\(encodedQuery)&type=\(typesParam)&limit=\(limit)"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-                throw SpotifyAPIError.invalidResponse
-            }
-
-            var tracks: [Track] = []
-            var albums: [Album] = []
-            var artists: [Artist] = []
-            var playlists: [Playlist] = []
-
-            // Parse tracks
-            if let tracksObj = json["tracks"] as? [String: Any],
-               let items = tracksObj["items"] as? [[String: Any]]
-            {
-                tracks = items.compactMap { item in
-                    guard let id = item["id"] as? String,
-                          let name = item["name"] as? String,
-                          let uri = item["uri"] as? String,
-                          let durationMs = item["duration_ms"] as? Int
-                    else {
-                        return nil
-                    }
-
-                    let artistsArray = item["artists"] as? [[String: Any]]
-                    let artistName = artistsArray?.first?["name"] as? String ?? "Unknown"
-                    let artistId = artistsArray?.first?["id"] as? String
-
-                    let albumData = item["album"] as? [String: Any]
-                    let albumName = albumData?["name"] as? String ?? ""
-                    let albumId = albumData?["id"] as? String
-                    let albumImages = albumData?["images"] as? [[String: Any]]
-                    let imageURLString = albumImages?.first?["url"] as? String
-                    let imageURL = imageURLString.flatMap { URL(string: $0) }
-
-                    let externalUrls = item["external_urls"] as? [String: Any]
-                    let externalUrl = externalUrls?["spotify"] as? String
-
-                    return Track(
-                        id: id,
-                        name: name,
-                        uri: uri,
-                        durationMs: durationMs,
-                        trackNumber: nil,
-                        externalUrl: externalUrl,
-                        albumId: albumId,
-                        artistId: artistId,
-                        artistName: artistName,
-                        albumName: albumName,
-                        imageURL: imageURL,
-                    )
-                }
-            }
-
-            // Parse albums
-            if let albumsObj = json["albums"] as? [String: Any],
-               let items = albumsObj["items"] as? [[String: Any]]
-            {
-                albums = items.compactMap { item in
-                    guard let id = item["id"] as? String,
-                          let name = item["name"] as? String,
-                          let uri = item["uri"] as? String
-                    else {
-                        return nil
-                    }
-
-                    let artistsArray = item["artists"] as? [[String: Any]]
-                    let artistName = artistsArray?.first?["name"] as? String ?? "Unknown"
-                    let artistId = artistsArray?.first?["id"] as? String
-                    let totalTracks = item["total_tracks"] as? Int ?? 0
-                    let releaseDate = item["release_date"] as? String ?? ""
-                    let albumType = item["album_type"] as? String
-                    let images = item["images"] as? [[String: Any]]
-                    let imageURLString = images?.first?["url"] as? String
-                    let imageURL = imageURLString.flatMap { URL(string: $0) }
-                    let externalUrls = item["external_urls"] as? [String: Any]
-                    let externalUrl = externalUrls?["spotify"] as? String
-
-                    return Album(
-                        id: id,
-                        name: name,
-                        uri: uri,
-                        imageURL: imageURL,
-                        releaseDate: releaseDate,
-                        albumType: albumType,
-                        externalUrl: externalUrl,
-                        artistId: artistId,
-                        artistName: artistName,
-                        knownTrackCount: totalTracks,
-                    )
-                }
-            }
-
-            // Parse artists
-            if let artistsObj = json["artists"] as? [String: Any],
-               let items = artistsObj["items"] as? [[String: Any]]
-            {
-                artists = items.compactMap { item in
-                    guard let id = item["id"] as? String,
-                          let name = item["name"] as? String,
-                          let uri = item["uri"] as? String
-                    else {
-                        return nil
-                    }
-
-                    let genres = item["genres"] as? [String] ?? []
-                    let images = item["images"] as? [[String: Any]]
-                    let imageURLString = images?.first?["url"] as? String
-                    let imageURL = imageURLString.flatMap { URL(string: $0) }
-                    let followers = (item["followers"] as? [String: Any])?["total"] as? Int
-
-                    return Artist(
-                        id: id,
-                        name: name,
-                        uri: uri,
-                        imageURL: imageURL,
-                        genres: genres,
-                        followers: followers,
-                    )
-                }
-            }
-
-            // Parse playlists
-            if let playlistsObj = json["playlists"] as? [String: Any],
-               let itemsArray = playlistsObj["items"] as? [Any]
-            {
-                // Filter out NSNull values and cast to dictionaries
-                let items = itemsArray.compactMap { $0 as? [String: Any] }
-
-                playlists = items.compactMap { item in
-                    guard let id = item["id"] as? String,
-                          let name = item["name"] as? String,
-                          let uri = item["uri"] as? String
-                    else {
-                        return nil
-                    }
-
-                    let description = item["description"] as? String
-                    let images = item["images"] as? [[String: Any]]
-                    let imageURLString = images?.first?["url"] as? String
-                    let imageURL = imageURLString.flatMap { URL(string: $0) }
-                    let trackCount = (item["tracks"] as? [String: Any])?["total"] as? Int ?? 0
-                    let owner = item["owner"] as? [String: Any]
-                    let ownerId = owner?["id"] as? String ?? ""
-                    let ownerName = owner?["display_name"] as? String ?? ownerId
-
-                    return Playlist(
-                        id: id,
-                        name: name,
-                        description: description,
-                        imageURL: imageURL,
-                        uri: uri,
-                        isPublic: true,
-                        ownerId: ownerId,
-                        ownerName: ownerName.isEmpty ? "Unknown" : ownerName,
-                        knownTrackCount: trackCount,
-                    )
-                }
-            }
-
-            return SearchResults(
-                tracks: tracks,
-                albums: albums,
-                artists: artists,
-                playlists: playlists,
-            )
-
-        case 401:
-            throw SpotifyAPIError.unauthorized
-
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    // MARK: - Recently Played
-
-    /// Fetches recently played tracks
-    static func fetchRecentlyPlayed(
-        accessToken: String,
-        limit: Int = 50,
-    ) async throws -> RecentlyPlayedResponse {
-        let urlString = "\(baseURL)/me/player/recently-played?limit=\(limit)"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-                throw SpotifyAPIError.invalidResponse
-            }
-
-            guard let itemsArray = json["items"] as? [[String: Any]] else {
-                throw SpotifyAPIError.invalidResponse
-            }
-
-            let items = itemsArray.compactMap { item -> RecentlyPlayedItem? in
-                guard let trackData = item["track"] as? [String: Any],
-                      let trackId = trackData["id"] as? String,
-                      let trackName = trackData["name"] as? String,
-                      let trackUri = trackData["uri"] as? String,
-                      let durationMs = trackData["duration_ms"] as? Int,
-                      let playedAt = item["played_at"] as? String
-                else {
-                    return nil
-                }
-
-                let artistsArray = trackData["artists"] as? [[String: Any]]
-                let artistName = artistsArray?.first?["name"] as? String ?? "Unknown"
-                let artistId = artistsArray?.first?["id"] as? String
-
-                let albumData = trackData["album"] as? [String: Any]
-                let albumName = albumData?["name"] as? String ?? ""
-                let albumId = albumData?["id"] as? String
-                let albumImages = albumData?["images"] as? [[String: Any]]
-                let imageURLString = albumImages?.first?["url"] as? String
-                let imageURL = imageURLString.flatMap { URL(string: $0) }
-
-                let externalUrls = trackData["external_urls"] as? [String: Any]
-                let externalUrl = externalUrls?["spotify"] as? String
-
-                let track = SearchTrack(
-                    id: trackId,
-                    name: trackName,
-                    uri: trackUri,
-                    artistName: artistName,
-                    albumName: albumName,
-                    imageURL: imageURL,
-                    durationMs: durationMs,
-                    albumId: albumId,
-                    artistId: artistId,
-                    externalUrl: externalUrl,
-                )
-
-                // Parse context if available
-                var context: PlaybackContext? = nil
-                if let contextData = item["context"] as? [String: Any],
-                   let contextType = contextData["type"] as? String,
-                   let contextUri = contextData["uri"] as? String
-                {
-                    context = PlaybackContext(type: contextType, uri: contextUri)
-                }
-
-                return RecentlyPlayedItem(
-                    id: playedAt, // Use playedAt as unique ID
-                    track: track,
-                    playedAt: playedAt,
-                    context: context,
-                )
-            }
-
-            return RecentlyPlayedResponse(items: items)
-
-        case 401:
-            throw SpotifyAPIError.unauthorized
-
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    // MARK: - Spotify Connect (Devices)
-
-    /// Fetches available Spotify Connect devices
-    static func fetchAvailableDevices(accessToken: String) async throws -> DevicesResponse {
-        let urlString = "\(baseURL)/me/player/devices"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-                  let devicesArray = json["devices"] as? [[String: Any]]
-            else {
-                throw SpotifyAPIError.invalidResponse
-            }
-
-            let devices = devicesArray.compactMap { deviceData -> SpotifyDevice? in
-                guard let id = deviceData["id"] as? String,
-                      let name = deviceData["name"] as? String,
-                      let type = deviceData["type"] as? String,
-                      let isActive = deviceData["is_active"] as? Bool,
-                      let isPrivateSession = deviceData["is_private_session"] as? Bool,
-                      let isRestricted = deviceData["is_restricted"] as? Bool
-                else {
-                    return nil
-                }
-
-                let volumePercent = deviceData["volume_percent"] as? Int
-
-                return SpotifyDevice(
-                    id: id,
-                    name: name,
-                    type: type,
-                    isActive: isActive,
-                    isPrivateSession: isPrivateSession,
-                    isRestricted: isRestricted,
-                    volumePercent: volumePercent,
-                )
-            }
-
-            return DevicesResponse(devices: devices)
-
-        case 401:
-            throw SpotifyAPIError.unauthorized
-
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    /// Transfers playback to a different device
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - deviceId: The device ID to transfer playback to
-    ///   - play: Whether to start playing on the new device (default: true)
-    static func transferPlayback(accessToken: String, deviceId: String, play: Bool = true) async throws {
-        let urlString = "\(baseURL)/me/player"
-        #if DEBUG
-            apiLogger.debug("[PUT] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-
-        // Create request body
-        let body: [String: Any] = [
-            "device_ids": [deviceId],
-            "play": play,
-        ]
-        request.httpBody = try? JSONSerialization.data(withJSONObject: body)
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 204:
-            // Success - playback transferred
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 403:
-            throw SpotifyAPIError.apiError("Device is restricted and cannot accept playback")
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    /// Starts playback of specific tracks on a device
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - deviceId: The device ID to play on
-    ///   - trackUris: Array of track URIs to play
-    ///   - positionMs: Position in the first track to start from (default: 0)
-    static func startPlayback(
-        accessToken: String,
-        deviceId: String,
-        trackUris: [String],
-        positionMs: Int = 0,
-    ) async throws {
-        var urlComponents = URLComponents(string: "\(baseURL)/me/player/play")!
-        urlComponents.queryItems = [URLQueryItem(name: "device_id", value: deviceId)]
-
-        guard let url = urlComponents.url else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        let urlString = url.absoluteString
-        #if DEBUG
-            apiLogger.debug("[PUT] \(urlString)")
-        #endif
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-
-        // Create request body with track URIs and position
-        let body: [String: Any] = [
-            "uris": trackUris,
-            "position_ms": positionMs,
-        ]
-        request.httpBody = try? JSONSerialization.data(withJSONObject: body)
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200, 204:
-            // Success - playback started
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 403:
-            throw SpotifyAPIError.apiError("Device is restricted and cannot accept playback")
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    // MARK: - Spotify Connect Playback Controls
-
-    /// Pauses playback on the active Spotify Connect device
-    static func pausePlayback(accessToken: String) async throws {
-        let urlString = "\(baseURL)/me/player/pause"
-        #if DEBUG
-            apiLogger.debug("[PUT] \(urlString)")
-        #endif
-
-        let url = URL(string: urlString)!
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (_, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse,
-              (200 ... 204).contains(httpResponse.statusCode) || httpResponse.statusCode == 403
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-    }
-
-    /// Resumes playback on the active Spotify Connect device
-    static func resumePlayback(accessToken: String) async throws {
-        let urlString = "\(baseURL)/me/player/play"
-        #if DEBUG
-            apiLogger.debug("[PUT] \(urlString)")
-        #endif
-
-        let url = URL(string: urlString)!
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (_, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse,
-              (200 ... 204).contains(httpResponse.statusCode) || httpResponse.statusCode == 403
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-    }
-
-    /// Skips to the next track on the active Spotify Connect device
-    static func skipToNext(accessToken: String) async throws {
-        let urlString = "\(baseURL)/me/player/next"
-        #if DEBUG
-            apiLogger.debug("[POST] \(urlString)")
-        #endif
-
-        let url = URL(string: urlString)!
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (_, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse,
-              (200 ... 204).contains(httpResponse.statusCode) || httpResponse.statusCode == 403
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-    }
-
-    /// Skips to the previous track on the active Spotify Connect device
-    static func skipToPrevious(accessToken: String) async throws {
-        let urlString = "\(baseURL)/me/player/previous"
-        #if DEBUG
-            apiLogger.debug("[POST] \(urlString)")
-        #endif
-
-        let url = URL(string: urlString)!
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (_, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse,
-              (200 ... 204).contains(httpResponse.statusCode) || httpResponse.statusCode == 403
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-    }
-
-    /// Seeks to a position in the current track on the active Spotify Connect device
-    static func seekToPosition(accessToken: String, positionMs: Int) async throws {
-        var urlComponents = URLComponents(string: "\(baseURL)/me/player/seek")!
-        urlComponents.queryItems = [URLQueryItem(name: "position_ms", value: String(positionMs))]
-
-        guard let url = urlComponents.url else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        let urlString = url.absoluteString
-        #if DEBUG
-            apiLogger.debug("[PUT] \(urlString)")
-        #endif
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (_, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse,
-              (200 ... 204).contains(httpResponse.statusCode) || httpResponse.statusCode == 403
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-    }
-
-    /// Sets the volume on the active Spotify Connect device
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - volumePercent: Volume level (0-100)
-    ///   - deviceId: Optional device ID to set volume on (uses active device if nil)
-    static func setVolume(accessToken: String, volumePercent: Int, deviceId: String? = nil) async throws {
-        var urlComponents = URLComponents(string: "\(baseURL)/me/player/volume")!
-        var queryItems = [URLQueryItem(name: "volume_percent", value: String(volumePercent))]
-        if let deviceId {
-            queryItems.append(URLQueryItem(name: "device_id", value: deviceId))
-        }
-        urlComponents.queryItems = queryItems
-
-        guard let url = urlComponents.url else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        let urlString = url.absoluteString
-        #if DEBUG
-            apiLogger.debug("[PUT] \(urlString)")
-        #endif
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200, 204:
-            // Success
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 403:
-            throw SpotifyAPIError.apiError("Volume control not available for this device")
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    /// Fetches current playback state including active device and track info
-    static func fetchPlaybackState(accessToken: String) async throws -> PlaybackState? {
-        let urlString = "\(baseURL)/me/player"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-                throw SpotifyAPIError.invalidResponse
-            }
-
-            // Parse device info
-            var device: SpotifyDevice?
-            if let deviceData = json["device"] as? [String: Any],
-               let id = deviceData["id"] as? String,
-               let name = deviceData["name"] as? String,
-               let type = deviceData["type"] as? String
-            {
-                device = SpotifyDevice(
-                    id: id,
-                    name: name,
-                    type: type,
-                    isActive: deviceData["is_active"] as? Bool ?? true,
-                    isPrivateSession: deviceData["is_private_session"] as? Bool ?? false,
-                    isRestricted: deviceData["is_restricted"] as? Bool ?? false,
-                    volumePercent: deviceData["volume_percent"] as? Int,
-                )
-            }
-
-            // Parse track info
-            let currentTrack: PlaybackTrack? = (json["item"] as? [String: Any]).flatMap { PlaybackTrack(from: $0) }
-
-            let isPlaying = json["is_playing"] as? Bool ?? false
-            let progressMs = json["progress_ms"] as? Int ?? 0
-            let shuffleState = json["shuffle_state"] as? Bool ?? false
-            let repeatState = json["repeat_state"] as? String ?? "off"
-
-            return PlaybackState(
-                device: device,
-                isPlaying: isPlaying,
-                progressMs: progressMs,
-                currentTrack: currentTrack,
-                shuffleState: shuffleState,
-                repeatState: repeatState,
-            )
-
-        case 204:
-            // No active playback
-            return nil
-
-        case 401:
-            throw SpotifyAPIError.unauthorized
-
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    /// Fetches the user's playback queue
-    static func fetchQueue(accessToken: String) async throws -> QueueResponse {
-        let urlString = "\(baseURL)/me/player/queue"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-                throw SpotifyAPIError.invalidResponse
-            }
-
-            // Parse currently playing track
-            let currentlyPlaying: QueueTrack? = (json["currently_playing"] as? [String: Any]).flatMap { QueueTrack(from: $0) }
-
-            // Parse queue
-            let queueArray = json["queue"] as? [[String: Any]] ?? []
-            let queue = queueArray.compactMap { QueueTrack(from: $0) }
-
-            return QueueResponse(currentlyPlaying: currentlyPlaying, queue: queue)
-
-        case 204:
-            // No active playback
-            return QueueResponse(currentlyPlaying: nil, queue: [])
-
-        case 401:
-            throw SpotifyAPIError.unauthorized
-
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    // MARK: - Playlist Management
-
-    /// Gets the current user's Spotify user ID
-    /// - Parameter accessToken: Spotify access token
-    /// - Returns: The user's Spotify ID
-    static func getCurrentUserId(accessToken: String) async throws -> String {
-        let urlString = "\(baseURL)/me"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-                  let userId = json["id"] as? String
-            else {
-                throw SpotifyAPIError.invalidResponse
-            }
-            return userId
-
-        case 401:
-            throw SpotifyAPIError.unauthorized
-
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    /// Creates a new playlist for the user
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - userId: The user's Spotify ID
-    ///   - name: The name for the new playlist
-    ///   - description: Optional description for the playlist
-    ///   - isPublic: Whether the playlist should be public (default: false)
-    /// - Returns: The created playlist
-    static func createPlaylist(
-        accessToken: String,
-        userId: String,
-        name: String,
-        description: String? = nil,
-        isPublic: Bool = false,
-    ) async throws -> APIPlaylist {
-        let urlString = "\(baseURL)/users/\(userId)/playlists"
-        #if DEBUG
-            apiLogger.debug("[POST] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-
-        let body: [String: Any] = [
-            "name": name,
-            "description": description ?? "",
-            "public": isPublic,
-        ]
-        request.httpBody = try? JSONSerialization.data(withJSONObject: body)
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200, 201:
-            guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-                  let id = json["id"] as? String,
-                  let name = json["name"] as? String,
-                  let uri = json["uri"] as? String,
-                  let owner = json["owner"] as? [String: Any],
-                  let ownerId = owner["id"] as? String
-            else {
-                throw SpotifyAPIError.invalidResponse
-            }
-
-            // Owner's display_name can be null, fall back to id
-            let ownerName = owner["display_name"] as? String ?? ownerId
-
-            let description = json["description"] as? String
-            let isPublic = json["public"] as? Bool ?? false
-
-            var imageURL: URL?
-            if let images = json["images"] as? [[String: Any]],
-               let firstImage = images.first,
-               let urlString = firstImage["url"] as? String
-            {
-                imageURL = URL(string: urlString)
-            }
-
-            return APIPlaylist(
-                id: id,
-                name: name,
-                uri: uri,
-                description: description,
-                imageURL: imageURL,
-                trackCount: 0,
-                ownerId: ownerId,
-                ownerName: ownerName,
-                isPublic: isPublic,
-                totalDurationMs: nil,
-            )
-
-        case 401:
-            throw SpotifyAPIError.unauthorized
-
-        case 403:
-            throw SpotifyAPIError.apiError("Not authorized to create playlists for this user")
-
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    /// Adds tracks to an existing playlist
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - playlistId: The ID of the playlist to add tracks to
-    ///   - trackUris: Array of Spotify track URIs (e.g., "spotify:track:xxx")
-    static func addTracksToPlaylist(
-        accessToken: String,
-        playlistId: String,
-        trackUris: [String],
-    ) async throws {
-        let urlString = "\(baseURL)/playlists/\(playlistId)/tracks"
-        #if DEBUG
-            apiLogger.debug("[POST] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-
-        let body: [String: Any] = ["uris": trackUris]
-        request.httpBody = try? JSONSerialization.data(withJSONObject: body)
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 201:
-            // Success - tracks added
-            break
-
-        case 401:
-            throw SpotifyAPIError.unauthorized
-
-        case 403:
-            throw SpotifyAPIError.apiError("Not authorized to modify this playlist")
-
-        case 404:
-            throw SpotifyAPIError.notFound
-
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    /// Updates playlist details (name and/or description)
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - playlistId: The ID of the playlist to update
-    ///   - newName: New name for the playlist (optional)
-    ///   - newDescription: New description for the playlist (optional)
-    static func updatePlaylistDetails(
-        accessToken: String,
-        playlistId: String,
-        newName: String? = nil,
-        newDescription: String? = nil,
-    ) async throws {
-        let urlString = "\(baseURL)/playlists/\(playlistId)"
-        #if DEBUG
-            apiLogger.debug("[PUT] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-
-        var body: [String: Any] = [:]
-        if let newName {
-            body["name"] = newName
-        }
-        if let newDescription {
-            body["description"] = newDescription
-        }
-        request.httpBody = try? JSONSerialization.data(withJSONObject: body)
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            break
-
-        case 401:
-            throw SpotifyAPIError.unauthorized
-
-        case 403:
-            throw SpotifyAPIError.apiError("Not authorized to modify this playlist")
-
-        case 404:
-            throw SpotifyAPIError.notFound
-
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    /// Deletes (unfollows) a playlist
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - playlistId: The ID of the playlist to delete
-    static func deletePlaylist(
-        accessToken: String,
-        playlistId: String,
-    ) async throws {
-        let urlString = "\(baseURL)/playlists/\(playlistId)/followers"
-        #if DEBUG
-            apiLogger.debug("[DELETE] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "DELETE"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            break
-
-        case 401:
-            throw SpotifyAPIError.unauthorized
-
-        case 403:
-            throw SpotifyAPIError.apiError("Not authorized to delete this playlist")
-
-        case 404:
-            throw SpotifyAPIError.notFound
-
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    /// Removes tracks from a playlist
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - playlistId: The ID of the playlist
-    ///   - trackUris: Array of track URIs to remove
-    static func removeTracksFromPlaylist(
-        accessToken: String,
-        playlistId: String,
-        trackUris: [String],
-    ) async throws {
-        let urlString = "\(baseURL)/playlists/\(playlistId)/tracks"
-        #if DEBUG
-            apiLogger.debug("[DELETE] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "DELETE"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-
-        // Format: { "tracks": [{ "uri": "spotify:track:xxx" }, ...] }
-        let tracks = trackUris.map { ["uri": $0] }
-        let body: [String: Any] = ["tracks": tracks]
-        request.httpBody = try? JSONSerialization.data(withJSONObject: body)
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            break
-
-        case 401:
-            throw SpotifyAPIError.unauthorized
-
-        case 403:
-            throw SpotifyAPIError.apiError("Not authorized to modify this playlist")
-
-        case 404:
-            throw SpotifyAPIError.notFound
-
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    /// Reorders tracks in a playlist
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - playlistId: The ID of the playlist
-    ///   - rangeStart: The position of the first track to be reordered
-    ///   - insertBefore: The position where the tracks should be inserted
-    ///   - rangeLength: Number of tracks to reorder (default 1)
-    static func reorderPlaylistTracks(
-        accessToken: String,
-        playlistId: String,
-        rangeStart: Int,
-        insertBefore: Int,
-        rangeLength: Int = 1,
-    ) async throws {
-        let urlString = "\(baseURL)/playlists/\(playlistId)/tracks"
-        #if DEBUG
-            apiLogger.debug("[PUT] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-
-        let body: [String: Any] = [
-            "range_start": rangeStart,
-            "insert_before": insertBefore,
-            "range_length": rangeLength,
-        ]
-        request.httpBody = try? JSONSerialization.data(withJSONObject: body)
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200, 201:
-            break
-
-        case 401:
-            throw SpotifyAPIError.unauthorized
-
-        case 403:
-            throw SpotifyAPIError.apiError("Not authorized to modify this playlist")
-
-        case 404:
-            throw SpotifyAPIError.notFound
-
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    /// Replaces all tracks in a playlist (used for reordering)
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - playlistId: The ID of the playlist
-    ///   - trackUris: Array of track URIs in the desired order
-    static func replacePlaylistTracks(
-        accessToken: String,
-        playlistId: String,
-        trackUris: [String],
-    ) async throws {
-        let urlString = "\(baseURL)/playlists/\(playlistId)/tracks"
-        #if DEBUG
-            apiLogger.debug("[PUT] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-
-        let body: [String: Any] = ["uris": trackUris]
-        request.httpBody = try? JSONSerialization.data(withJSONObject: body)
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200, 201:
-            break
-
-        case 401:
-            throw SpotifyAPIError.unauthorized
-
-        case 403:
-            throw SpotifyAPIError.apiError("Not authorized to modify this playlist")
-
-        case 404:
-            throw SpotifyAPIError.notFound
-
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-    }
-
-    // MARK: - User Top Items
-
-    /// Fetches user's top artists from Spotify Web API
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - timeRange: Time range for calculating top items
-    ///   - limit: Number of artists to fetch (max 50)
-    ///   - offset: Offset for pagination
-    static func fetchUserTopArtists(
-        accessToken: String,
-        timeRange: TopItemsTimeRange = .mediumTerm,
-        limit: Int = 20,
-        offset: Int = 0,
-    ) async throws -> TopArtistsResponse {
-        let urlString = "\(baseURL)/me/top/artists?time_range=\(timeRange.rawValue)&limit=\(limit)&offset=\(offset)"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Accept")
-
-        let (data, response): (Data, URLResponse)
-        do {
-            (data, response) = try await URLSession.shared.data(for: request)
-        } catch {
-            throw SpotifyAPIError.networkError(error)
-        }
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-
-        // Parse JSON response
-        guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-              let items = json["items"] as? [[String: Any]],
-              let total = json["total"] as? Int
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        let artists = items.compactMap { item -> APIArtist? in
-            guard let id = item["id"] as? String,
-                  let name = item["name"] as? String,
-                  let uri = item["uri"] as? String
-            else {
-                return nil
-            }
-
-            let genres = (item["genres"] as? [String]) ?? []
-            let followersDict = item["followers"] as? [String: Any]
-            let followers = followersDict?["total"] as? Int ?? 0
-
-            var imageURL: URL?
-            if let images = item["images"] as? [[String: Any]],
-               let firstImage = images.first,
-               let urlString = firstImage["url"] as? String
-            {
-                imageURL = URL(string: urlString)
-            }
-
-            return APIArtist(
-                id: id,
-                name: name,
-                uri: uri,
-                imageURL: imageURL,
-                genres: genres,
-                followers: followers
-            )
-        }
-
-        let nextOffset = offset + limit
-        let hasMore = nextOffset < total
-
-        return TopArtistsResponse(
-            artists: artists,
-            total: total,
-            hasMore: hasMore,
-            nextOffset: hasMore ? nextOffset : nil,
-        )
-    }
-
-    // MARK: - Browse
-
-    /// Fetches new album releases from Spotify Web API
-    /// - Parameters:
-    ///   - accessToken: Spotify access token
-    ///   - limit: Number of albums to fetch (max 50)
-    ///   - offset: Offset for pagination
-    static func fetchNewReleases(
-        accessToken: String,
-        limit: Int = 20,
-        offset: Int = 0,
-    ) async throws -> NewReleasesResponse {
-        let urlString = "\(baseURL)/browse/new-releases?limit=\(limit)&offset=\(offset)"
-        #if DEBUG
-            apiLogger.debug("[GET] \(urlString)")
-        #endif
-
-        guard let url = URL(string: urlString) else {
-            throw SpotifyAPIError.invalidURI
-        }
-
-        var request = URLRequest(url: url)
-        request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Accept")
-
-        let (data, response): (Data, URLResponse)
-        do {
-            (data, response) = try await URLSession.shared.data(for: request)
-        } catch {
-            throw SpotifyAPIError.networkError(error)
-        }
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        switch httpResponse.statusCode {
-        case 200:
-            break
-        case 401:
-            throw SpotifyAPIError.unauthorized
-        case 404:
-            throw SpotifyAPIError.notFound
-        default:
-            if let errorJson = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-               let error = errorJson["error"] as? [String: Any],
-               let message = error["message"] as? String
-            {
-                throw SpotifyAPIError.apiError(message)
-            }
-            throw SpotifyAPIError.apiError("HTTP \(httpResponse.statusCode)")
-        }
-
-        // Parse JSON response
-        guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-              let albumsContainer = json["albums"] as? [String: Any],
-              let items = albumsContainer["items"] as? [[String: Any]],
-              let total = albumsContainer["total"] as? Int
-        else {
-            throw SpotifyAPIError.invalidResponse
-        }
-
-        let albums = items.compactMap { item -> APIAlbum? in
-            guard let id = item["id"] as? String,
-                  let name = item["name"] as? String,
-                  let uri = item["uri"] as? String
-            else {
-                return nil
-            }
-
-            // Extract artist info
-            var artistName = "Unknown Artist"
-            if let artists = item["artists"] as? [[String: Any]],
-               let firstArtist = artists.first
-            {
-                artistName = (firstArtist["name"] as? String) ?? artistName
-            }
-
-            // Extract image
-            var imageURL: URL?
-            if let images = item["images"] as? [[String: Any]],
-               let firstImage = images.first,
-               let urlString = firstImage["url"] as? String
-            {
-                imageURL = URL(string: urlString)
-            }
-
-            let trackCount = item["total_tracks"] as? Int ?? 0
-            let releaseDate = (item["release_date"] as? String) ?? ""
-            let albumType = (item["album_type"] as? String) ?? "album"
-
-            return APIAlbum(
-                id: id,
-                name: name,
-                uri: uri,
-                artistName: artistName,
-                artistId: nil,
-                imageURL: imageURL,
-                trackCount: trackCount,
-                releaseDate: releaseDate,
-                albumType: albumType,
-                totalDurationMs: nil,
-                externalUrl: nil,
-            )
-        }
-
-        let nextOffset = offset + limit
-        let hasMore = nextOffset < total
-
-        return NewReleasesResponse(
-            albums: albums,
-            total: total,
-            hasMore: hasMore,
-            nextOffset: hasMore ? nextOffset : nil,
-        )
-    }
-}
